misc fixes and moved utility funcs to global utility module
